Add Initial Authentication, Models/Api Endpoints & JsonApi Middleware
This commit is a little jumbled up because it was worked on before the
contribution guide. It adds authentication using Laravel's Passport, the
first few models (Users, Podcasts, Contributors), the start of the
endpoints for Podcasts and the start of a JsonApi middleware. It also
adds an IDE helper for PHP Storm.