Removed '/api' Prefix From Api Routes
Now the api routes will only be prefixed with '/v1'. This change has made because the server hosting this code will only be serving up the API and nothing else. So there is no need to prefix the api.

Also unmapped the web routes, so any non api requests will return a 404 Not Found.